requirements.txt
appnope==0.1.2
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bandit==1.7.1
decorator==5.1.0
distro==1.6.0
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.24
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
numpy==1.21.5
pandas==1.3.5
parso==0.8.3
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.1
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
tabula==1.0.5
tabula-py==2.3.0
toml==0.10.2
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
